Break tests into separate files (create one for MR)
